requirements.txt
nltk==3.8.1
numpy==1.26.4
pandas==2.2.0
Requests==2.31.0
scikit_learn==1.4.0
tensorflow_cpu==2.15.0.post1
